Remove checking "resolving" in  needToResolveOnBrowser
The resolving condition will cause a bug when the component is resolving, then the current component is unmounted and then the component is recreated again right before the module is resolved. 
In this case, the component should get the existing resolver and update the rendering after the module is resolved. However, the resolving checking blocks this update.